requirements.txt
django
django-ckeditor
googletrans
djangorestframework
django-redis
# ckeditor_uploader
httpcore==0.12.0 